+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## devel
 
+* add cluster alloctable memory and cpu to summary
 * add cluster memory and cpu capacity to summary
 * add cluster version to summary